Start refactoring derives for compare impls
