+++ rkyv/Cargo.toml
@@ -16,7 +16,7 @@ readme = "crates-io.md"
 [dependencies]
 bytecheck = { version = "0.4", optional = true }
 memoffset = "0.6"
-ptr_meta = { version = "0.1.1" }
+ptr_meta = { version = "0.1" }
 rkyv_derive = { version = "=0.4.0", path = "../rkyv_derive" }
 seahash = { version = "4.0", optional = true }
