+++ rkyv_dyn/Cargo.toml
@@ -17,7 +17,7 @@ readme = "crates-io.md"
 bytecheck = { version = "0.4", optional = true }
 inventory = "0.1"
 lazy_static = "1.4"
-ptr_meta = { version = "0.1.1" }
+ptr_meta = { version = "0.1" }
 rkyv = { version = "0.4", path = "../rkyv" }
 rkyv_dyn_derive = { version = "=0.4.0", path = "../rkyv_dyn_derive" }
 rkyv_typename = { version = "0.4", path = "../rkyv_typename" }
